# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "Documentação da API do sistema",
    "contact": {
      "name": "Equipe de Desenvolvimento",
      "email": "dev@example.com"
    }
  },
  "host": "localhost:3000",
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desenvolvimento"
    }
  ],
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/disciplina/{deletarDisciplinacodigo}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "deletarDisciplinacodigo",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/disciplina/{atulizarDisciplinacodigo}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "atulizarDisciplinacodigo",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}